+++ cli/cli/commands/enclave/inspect/user_services.go
@@ -13,6 +13,7 @@ import (
 
 const (
 	userServiceGUIDColHeader                                    = "GUID"
+	userServiceIDColHeader                                      = "ID"
 	userServiceHostMachinePortBindingsColHeader                 = "LocalPortBindings"
 	shouldShowStoppedContainersWhenGettingUserServiceContainers = true
 )
@@ -25,36 +26,42 @@ func printUserServices(ctx context.Context, dockerManager *docker_manager.Docker
 		return stacktrace.Propagate(err, "An error occurred getting user service containers by labels: '%+v'", userServiceLabels)
 	}
 
-	tablePrinter := output_printers.NewTablePrinter(userServiceGUIDColHeader, userServiceHostMachinePortBindingsColHeader)
+	tablePrinter := output_printers.NewTablePrinter(userServiceGUIDColHeader, userServiceIDColHeader, userServiceHostMachinePortBindingsColHeader)
 	sortedContainers, err := sortContainersByGUID(containers)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred sorting user service containers by GUID")
 	}
 	for _, container := range sortedContainers {
-		containerGuid, found := container.GetLabels()[schema.GUIDLabel]
+		serviceGuid, found := container.GetLabels()[schema.GUIDLabel]
 		if !found {
 			return stacktrace.NewError("No '%v' container label was found in container ID '%v' with labels '%+v'", schema.GUIDLabel, container.GetId(), container.GetLabels())
 		}
-		hostPortBindingsStrings := getContainerHostPortBindingStrings(container)
+		serviceId, found := container.GetLabels()[schema.IDLabel]
+		if !found {
+			return stacktrace.NewError("No '%v' container label was found in container ID '%v' with labels '%+v'", schema.IDLabel, container.GetId(), container.GetLabels())
+		}
 
+		hostPortBindingsStrings := getContainerHostPortBindingStrings(container)
 		firstHostPortBindingStr := ""
 		if hostPortBindingsStrings != nil {
 			firstHostPortBindingStr = hostPortBindingsStrings[0]
 			hostPortBindingsStrings = hostPortBindingsStrings[1:]
 		}
-		if err := tablePrinter.AddRow(containerGuid, firstHostPortBindingStr); err != nil {
-			return stacktrace.NewError(
+		if err := tablePrinter.AddRow(serviceGuid, serviceId, firstHostPortBindingStr); err != nil {
+			return stacktrace.Propagate(
+				err,
 				"An error occurred adding row for user service container '%v' to the table printer",
-				containerGuid,
+				serviceGuid,
 			)
 		}
 
 		for _, additionalHostPortBindingStr := range hostPortBindingsStrings {
-			if err := tablePrinter.AddRow("", additionalHostPortBindingStr); err != nil {
-				return stacktrace.NewError(
+			if err := tablePrinter.AddRow("", "", additionalHostPortBindingStr); err != nil {
+				return stacktrace.Propagate(
+					err,
 					"An error occurred adding additional host port binding '%v' row for user service container '%v' to the table printer",
 					additionalHostPortBindingStr,
-					containerGuid,
+					serviceGuid,
 				)
 			}
 		}